Merge MCTeamPotato's fix

+++ src/main/java/mcjty/lostcities/worldgen/lost/DamageArea.java
@@ -177,7 +177,7 @@ public class DamageArea {
     public float getDamageFactor() {
         float damage = 0.0f;
         for (Explosion explosion : explosions) {
-            double sq = explosion.getCenter().distSqr(chunkX * 16.0, (double) explosion.getCenter().getY(), chunkZ * 16.0, true);
+            double sq = explosion.getCenter().distSqr(chunkX * 16.0, explosion.getCenter().getY(), chunkZ * 16.0, true);
             if (sq < explosion.getSqradius()) {
                 double d = Math.sqrt(sq);
                 damage += 3.0f * (explosion.getRadius() - d) / explosion.getRadius();
